refactor: change the type from immer & export some function in a new file
